[quant][fx] Make scale, zero_point buffers in the model and use FQN (for quantized ops) Summary:
Currently scale and zero_point values are stored as constant values in the graph.
This prevents these values from being updated in the graph and also does not enable saving
these values to state_dict

After this PR we store scale/zero_point values for quantized ops as buffers in the root module
and createe get_attr nodes for them in the graph.
The scale and zero_point values will now be stored as part of the model state_dict of the FX quantized model.
Before this PR, we were not storing any qparams in state_dict so saving/loading a quantized fx model didn't work.

We also use the FQN of the module where the quantized ops are present to name these attributes so
that they can be uniquely  identified and mapped to quantized ops.

Test Plan:
python test/test_quantization.py TestQuantizeFx.test_qparams_buffers

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: eb9f9859790ac016c48ad5717d649804f1ce8090
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51166